pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "django-keycloak-admin"
version = "0.0.4"
authors = [
  { name="Erik Polzin", email="eriktpol@gmail.com" },
]
description = "Keycloak admin and REST auth."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    'python-keycloak>=4.0.0',
    'Django>=5.0.0',
    'djangorestframework>=3.0.0',
]

[project.optional-dependencies]
build = ["build", "twine"]
dev = ["pip-tools", "bumpver", "pytest", "pytest-django"]

[project.urls]
Homepage = "https://github.com/ErikPolzin/django-keycloak"
Issues = "https://github.com/ErikPolzin/django-keycloak/issues"

[tool.bumpver]
current_version = "0.0.4"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message = "bump version {old_version} -> {new_version}"
tag_message = "{new_version}"
tag_scope = "default"
pre_commit_hook = ""
post_commit_hook = ""
commit = true
tag = true
push = false

[tool.bumpver.file_patterns]
"pyproject.toml" = [
    'current_version = "{version}"',
    'version = "{version}"',
]
